added background region, SRDs now use mask, malloc works

+++ mpu/src/utilities.h
@@ -14,7 +14,7 @@
 void emb_vprintf(const char * str, va_list vaArgP);
 void emb_printf(const char * str, ...);
 
-uint32_t emb_log2(uint32_t value);
+uint8_t emb_log2(uint64_t value);
 uint32_t emb_pow2(int8_t value);
 
 #endif /* UTILITIES_H_ */
